// queries/types/RolesQuery.d.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL query operation: RolesQuery
// ====================================================

export interface RolesQuery_roles {
  __typename: "Role";
  roleId: string;
  name: string;
}

export interface RolesQuery {
  roles: RolesQuery_roles[];
}